requirements.txt
Babel==2.0
Jinja2==2.8
linecache2==1.0.0
MarkupSafe==0.23
passlib==1.6.5
Pillow==2.9.0
psycopg2==2.6.1
pytz==2015.4
PyYAML==3.11
requests==2.7.0
six==1.9.0
SQLAlchemy==1.0.8
traceback2==1.4.0
unittest2==1.1.0
Werkzeug==0.10.4
wheel==0.24.0
attrdict==2.0.0